requirements.txt
telethon
python-dotenv
pandas
sqlalchemy
psycopg2-binary
loguru
psycopg2-binary
 sqlalchemy 
 pandas 
 dbt-postgres